Final LoginReader structure
